feat: Implement core game structure with dedicated modules for constants, level generation, and sprite classes for leaves, platforms, and the player, alongside new assets.
