+++ constants.py
@@ -1,0 +1,25 @@
+
+import pygame
+
+# ------------------ Constants -------------------
+SCREEN_WIDTH, SCREEN_HEIGHT = 600, 1200
+FPS = 60
+
+# Colors
+BLACK = (0, 0, 0)
+SKY_BLUE = (145, 220, 255)
+WHITE = (255, 255, 255)
+
+# Fall Colors for Leaves
+FALL_COLORS = [
+    (139, 37, 0),   # Dark Red
+    (205, 55, 0),   # Burnt Orange
+    (238, 118, 0),  # Golden Orange
+    (205, 133, 63), # Peru Brown
+    (160, 82, 45),  # Sienna
+    (218, 165, 32), # Golden Rod
+    (184, 134, 11), # Dark Golden Rod
+    (139, 69, 19),  # Saddle Brown
+]
+
+MAX_LEAVES = 50
